# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Acro Game - API",
        "contact": {
            "name": "Vilius Grižas",
            "email": "viliusgrizas@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logs in the user, returns api token on success",
                "operationId": "App\\Http\\Controllers\\AuthenticationController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registration. Returns user info",
                "operationId": "App\\Http\\Controllers\\AuthenticationController::register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/poses": {
            "get": {
                "tags": [
                    "Poses"
                ],
                "summary": "Logs in the user, returns api token on success",
                "operationId": "App\\Http\\Controllers\\PoseController::index",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Generated user token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "uIPdxJeJiFyUnUnkhVE7eAgOS6EADpL7nEaWrsGl4rPENRAuyfrrMkBns0ndqEhlmWlTfSxl5Dnbfaf0ViYQ8b1vnbch"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoseIndexResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Poses"
                ],
                "summary": "Upload a pose to database. Will automatically have status 2 - Pending",
                "operationId": "App\\Http\\Controllers\\PoseController::post",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Generated user token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "uIPdxJeJiFyUnUnkhVE7eAgOS6EADpL7nEaWrsGl4rPENRAuyfrrMkBns0ndqEhlmWlTfSxl5Dnbfaf0ViYQ8b1vnbch"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PoseStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PosePostResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/poses/pending": {
            "get": {
                "tags": [
                    "Poses"
                ],
                "summary": "All the pending poses. Only for admins.",
                "operationId": "App\\Http\\Controllers\\PoseController::getPendingPoses",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Generated user token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "uIPdxJeJiFyUnUnkhVE7eAgOS6EADpL7nEaWrsGl4rPENRAuyfrrMkBns0ndqEhlmWlTfSxl5Dnbfaf0ViYQ8b1vnbch"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetPendingPosesResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/poses/{id}": {
            "get": {
                "tags": [
                    "Poses"
                ],
                "summary": "Get a specific pose by ID",
                "operationId": "App\\Http\\Controllers\\PoseController::show",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Generated user token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "uIPdxJeJiFyUnUnkhVE7eAgOS6EADpL7nEaWrsGl4rPENRAuyfrrMkBns0ndqEhlmWlTfSxl5Dnbfaf0ViYQ8b1vnbch"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoseShowResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Poses"
                ],
                "summary": "Updates a pose (can be used for enabling/changing status/disabling (functions as deleting)",
                "operationId": "App\\Http\\Controllers\\PoseController::update",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Generated user token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "uIPdxJeJiFyUnUnkhVE7eAgOS6EADpL7nEaWrsGl4rPENRAuyfrrMkBns0ndqEhlmWlTfSxl5Dnbfaf0ViYQ8b1vnbch"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PoseUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoseUpdateResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/poses/search": {
            "get": {
                "tags": [
                    "Poses"
                ],
                "summary": "Searches the database for poses by name. Minimum length of search term is 3 letters.",
                "operationId": "App\\Http\\Controllers\\PoseController::search",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Generated user token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "uIPdxJeJiFyUnUnkhVE7eAgOS6EADpL7nEaWrsGl4rPENRAuyfrrMkBns0ndqEhlmWlTfSxl5Dnbfaf0ViYQ8b1vnbch"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoseSearchResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "dupadupa666"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginRequest"
                }
            },
            "RegisterRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "dupadupa666"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RegisterRequest"
                }
            },
            "PoseStoreRequest": {
                "properties": {
                    "name": {
                        "required": true,
                        "type": "string",
                        "example": "Reverse Star"
                    },
                    "difficulty": {
                        "type": "ingeger",
                        "example": 1
                    },
                    "type": {
                        "type": "ingeger",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "example": "www.mickevičius.lt  "
                    },
                    "people_count": {
                        "type": "integer",
                        "example": "2"
                    },
                    "description": {
                        "type": "string",
                        "example": "Point!"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PoseStoreRequest"
                }
            },
            "PoseUpdateRequest": {
                "properties": {
                    "name": {
                        "required": true,
                        "type": "string",
                        "example": "Reverse Star"
                    },
                    "difficulty": {
                        "type": "ingeger",
                        "example": 1
                    },
                    "type": {
                        "type": "ingeger",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "example": "www.mickevičius.lt  "
                    },
                    "people_count": {
                        "type": "integer",
                        "example": "2"
                    },
                    "description": {
                        "type": "string",
                        "example": "Point!"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PoseStoreRequest"
                }
            },
...
            "LoginResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "token": {
                                "type": "string",
                                "example": "avdsfgy6v34etbrteav"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginResponse"
                }
            },
            "RegisterResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "2"
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "email": {
                                "type": "string",
                                "example": "user@email.com"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RegisterResponse"
                }
            },
            "GetPendingPosesResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Bob"
                                },
                                "difficulty": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "type": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "image_url": {
                                    "type": "string",
                                    "example": "www.best-website.com"
                                },
                                "people_count": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Point! "
                                },
                                "status": {
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "GetPendingPosesResponse"
                }
            },
            "PoseIndexResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Bob"
                                },
                                "difficulty": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "type": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "image_url": {
                                    "type": "string",
                                    "example": "www.best-website.com"
                                },
                                "people_count": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Point! "
                                },
                                "status": {
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PoseIndexResponse"
                }
            },
            "PosePostResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "2"
                            },
                            "name": {
                                "type": "string",
                                "example": "Bob"
                            },
                            "difficulty": {
                                "type": "integer",
                                "example": "2"
                            },
                            "type": {
                                "type": "integer",
                                "example": "2"
                            },
                            "image_url": {
                                "type": "string",
                                "example": "www.best-website.com"
                            },
                            "people_count": {
                                "type": "integer",
                                "example": "2"
                            },
                            "description": {
                                "type": "string",
                                "example": "Point! "
                            },
                            "status": {
                                "type": "integer",
                                "example": "1"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PosePostResponse"
                }
            },
            "PoseSearchResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Bob"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PoseSearchResponse"
                }
            },
            "PoseShowResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "2"
                            },
                            "name": {
                                "type": "string",
                                "example": "Bob"
                            },
                            "difficulty": {
                                "type": "integer",
                                "example": "2"
                            },
                            "type": {
                                "type": "integer",
                                "example": "2"
                            },
                            "image_url": {
                                "type": "string",
                                "example": "www.best-website.com"
                            },
                            "people_count": {
                                "type": "integer",
                                "example": "2"
                            },
                            "description": {
                                "type": "string",
                                "example": "Point! "
                            },
                            "status": {
                                "type": "integer",
                                "example": "1"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PoseShowResponse"
                }
            },
            "PoseUpdateResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "2"
                            },
                            "name": {
                                "type": "string",
                                "example": "Bob"
                            },
                            "difficulty": {
                                "type": "integer",
                                "example": "2"
                            },
                            "type": {
                                "type": "integer",
                                "example": "2"
                            },
                            "image_url": {
                                "type": "string",
                                "example": "www.best-website.com"
                            },
                            "people_count": {
                                "type": "integer",
                                "example": "2"
                            },
                            "description": {
                                "type": "string",
                                "example": "Point! "
                            },
                            "status": {
                                "type": "integer",
                                "example": "1"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PoseUpdateResponse"
                }
            },
...
        }
    }
}